requirements.txt
beautifulsoup4
sentence_splitter
requests
tqdm
pyyaml